improve: external scanner incremental reuse safety and observability
- Add IncrementalReuseExternalScanner interface with SupportsIncrementalReuse() opt-in for safe subtree reuse during incremental parses
- Zero arena nodes and child slices on allocation to prevent stale pointer leakage between parse passes
- Refactor parser retry logic into dedicated parser_retry.go for better code organization
- Extend incremental parse profiling with detailed reuse rejection counters and memory allocation metrics
- Fix LexState Default/EOF and NextState fields to use int32 for consistency
- Track field sources explicitly during reduce actions instead of post-hoc assignment
- Add comprehensive tests for external scanner incremental reuse policies and reset behavior
- Implement snippet parser pooling for efficient recovery parser reuse

+++ cmd/ts2go/embedded.go
@@ -232,7 +232,7 @@ func convertLexStates(src []LexStateEntry) []gotreesitter.LexState {
 			AcceptToken: gotreesitter.Symbol(s.Accept),
 			Skip:        false,
 			Default:     -1,
-			EOF:         s.EOF,
+			EOF:         int32(s.EOF),
 		}
 		if len(s.Transitions) > 0 {
 			state.Transitions = make([]gotreesitter.LexTransition, len(s.Transitions))
@@ -240,7 +240,7 @@ func convertLexStates(src []LexStateEntry) []gotreesitter.LexState {
 				state.Transitions[j] = gotreesitter.LexTransition{
 					Lo:        t.Lo,
 					Hi:        t.Hi,
-					NextState: t.Next,
+					NextState: int32(t.Next),
 					Skip:      t.Skip,
 				}
 			}
